models/paddlecv/requirements.txt
pytest
pytest-assume
pytest-html
pytest-timeout
pytest-repeat
allure-pytest
